+++ bin/varnishd/mgt/mgt_vcl.c
@@ -265,7 +265,8 @@ mgt_vcl_del(struct vclprog *vp)
 
 	VTAILQ_REMOVE(&vclhead, vp, list);
 	if (vp->fname != NULL) {
-		AZ(unlink(vp->fname));
+		if (!MGT_DO_DEBUG(DBG_VCL_KEEP))
+			AZ(unlink(vp->fname));
 		p = strrchr(vp->fname, '/');
 		AN(p);
 		*p = '\0';
